apps/recommendation-service-python/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
tensorflow-cpu==2.15.0
python-dotenv==1.0.1
numpy==1.26.3
